object-file.c: use size_t for header lengths
Continue walking the code path for the >4GB `hash-object --literally`
test. The `hash_object_file_literally()` function internally uses both
`hash_object_file()` and `write_object_file_prepare()`. Both function
signatures use `unsigned long` rather than `size_t` for the mem buffer
sizes. Use `size_t` instead, for LLP64 compatibility.

While at it, convert those function's object's header buffer length to
`size_t` for consistency. The value is already upcast to `uintmax_t` for
print format compatibility.

Note: The hash-object test still does not pass. A subsequent commit
continues to walk the call tree's lower level hash functions to identify
further fixes.

+++ object-store.h
@@ -247,7 +247,7 @@ int oid_object_info(struct repository *r, const struct object_id *, unsigned lon
 
 
 void hash_object_file(const struct git_hash_algo *algo, const void *buf,
-		      unsigned long len, enum object_type type,
+		      size_t len, enum object_type type,
 		      struct object_id *oid);
 
 /**
@@ -260,11 +260,11 @@ void hash_object_file(const struct git_hash_algo *algo, const void *buf,
  * was.
  */
 int hash_object_file_oideq(const struct git_hash_algo *algo, const void *buf,
-			   unsigned long len, enum object_type type,
+			   size_t len, enum object_type type,
 			   const struct object_id *oid,
 			   struct object_id *real_oidp);
 
-int write_object_file_flags(const void *buf, unsigned long len,
+int write_object_file_flags(const void *buf, size_t len,
 			    enum object_type type, struct object_id *oid,
 			    unsigned flags);
 static inline int write_object_file(const void *buf, unsigned long len,
